+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 * Spark: Support latest Spark 3.3.1 [`#1183`](https://github.com/OpenLineage/OpenLineage/pull/1183) [@pawel-big-lebowski](https://github.com/pawel-big-lebowski)
 * Python: add facets implementation to Python client [`#1233`](https://github.com/OpenLineage/OpenLineage/pull/1233) [@pawel-big-lebowski](https://github.com/pawel-big-lebowski)
+* Spark: Disable specified facets [`#1271`](https://github.com/OpenLineage/OpenLineage/pull/1271) [@pawel-big-lebowski](https://github.com/pawel-big-lebowski)
+
+    *Adds the ability to disable specified facets from generated OpenLineage events.*
 
 ### Removed
 * Deprecate HttpTransport.Builder in favor of HttpConfig [@collado-mike](https://github.com/collado-mike)